refactor(transaction): ADR-0120 Phase 5+6 — retire PaymentSagaOrchestrator, Temporal always-on
- Delete PaymentSagaOrchestrator, PaymentSaga, PaymentSagaEntity,
  PanachePaymentSagaRepository and their tests; Temporal is now the sole
  orchestrator (no feature flag).
- Extract SagaState enum to domain/saga so PaymentWorkflow.execute()
  return type survives the saga deletion.
- Remove TemporalConfig.enabled(); PaymentWorkerRegistrar.onStart()
  registers the worker unconditionally at boot.
- Add Flyway V14 tombstone: payment_sagas table marked RETIRED
  (write path closed, historical rows preserved for audit window).
- Update TransactionServiceTest: replace saga mocks with Temporal
  workflow stub; delete flag-gating tests.
- Bump version.txt 1.12.2 → 1.13.0 (minor: behavioral change).
- Mark ADR-0120 Delivery-Status: Shipped; update delivery log.
