+++ openbank-transaction-service/src/main/kotlin/com/openbank/transaction/application/workflow/PaymentWorkerRegistrar.kt
@@ -13,9 +13,7 @@ import jakarta.enterprise.event.Observes
 import org.jboss.logging.Logger
 
 /**
- * Registers the Temporal payment worker at boot — but only when the ADR-0120 P1 flag is on. With the
- * flag disabled (the default) this returns immediately and never touches the [WorkflowClient], so no
- * Temporal frontend dial happens at startup.
+ * Registers the Temporal payment worker at boot (ADR-0120 Phase 5: always-on after saga retirement).
  */
 @ApplicationScoped
 class PaymentWorkerRegistrar(
@@ -28,13 +26,6 @@ class PaymentWorkerRegistrar(
 
     @Suppress("UnusedParameter")
     fun onStart(@Observes event: StartupEvent) {
-        if (!temporalConfig.enabled()) {
-            log.info(
-                "Temporal payment orchestration disabled " +
-                    "(openbank.transaction.orchestration.temporal.enabled=false); skipping registration",
-            )
-            return
-        }
         log.infof("Registering Temporal payment worker on task queue '%s'", temporalConfig.taskQueue())
         val factory = WorkerFactory.newInstance(workflowClient)
         val worker = factory.newWorker(temporalConfig.taskQueue())
