+++ openbank-transaction-service/version.txt
@@ -1,1 +1,1 @@
-1.12.2
+1.13.0
